# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pannaga API",
    "description": "Hyperlocal Monsoon Onset & Break Prediction System for India",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/request-otp": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Request OTP for phone login",
        "operationId": "request_otp_api_v1_auth_request_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestOTPIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Request Otp Api V1 Auth Request Otp Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/verify-otp": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify OTP and issue JWT tokens",
        "operationId": "verify_otp_api_v1_auth_verify_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOTPIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Verify Otp Api V1 Auth Verify Otp Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/onboarding": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Complete farmer onboarding with consent",
        "operationId": "onboarding_api_v1_auth_onboarding_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Onboarding Api V1 Auth Onboarding Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IrrigationSource": {
        "type": "string",
        "enum": [
          "RAINFED",
          "CANAL",
          "BOREWELL_WELL",
          "TANK_POND",
          "DRIP_SPRINKLER",
          "OTHER",
          "BOREWELL",
          "TANK"
        ],
        "title": "IrrigationSource"
      },
      "OnboardingIn": {
        "properties": {
          "panchayat_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Panchayat Id"
          },
          "crop_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crop Type"
          },
          "irrigation_source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IrrigationSource"
              },
              {
                "type": "null"
              }
            ]
          },
          "farm_size_acres": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Farm Size Acres"
          },
          "soil_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Soil Type"
          },
          "consent": {
            "type": "boolean",
            "title": "Consent",
            "default": false
          },
          "preferred_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Language"
          }
        },
        "type": "object",
        "title": "OnboardingIn"
      },
...
      "RequestOTPIn": {
        "properties": {
          "phone_number": {
            "type": "string",
            "title": "Phone Number"
          }
        },
        "type": "object",
        "required": [
          "phone_number"
        ],
        "title": "RequestOTPIn"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyOTPIn": {
        "properties": {
          "phone_number": {
            "type": "string",
            "title": "Phone Number"
          },
          "otp": {
            "type": "string",
            "title": "Otp"
          }
        },
        "type": "object",
        "required": [
          "phone_number",
          "otp"
        ],
        "title": "VerifyOTPIn"
      }
    }
  }
}